Create new round logic
The code is now able to get the two <p> elements inside the scores div
and handle them as an array so the score can be displayed correctly,
this approach also optimizes the code by using only one variable
to get both <p> elements.

When the buttons are clicked, the playRound() function is called,
passing only the human score as an argument.